finance_system_v1/requirement.txt
flask_mysqldb
flask
python-dotenv